backend/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "uvicorn[standard]>=0.30.0",
    "icalendar>=6.3.2",
    "pydantic>=2.12.5",
    "pytz>=2025.2",
    "openai>=1.0.0",
    "parsedatetime>=2.6",
    "google-auth>=2.45.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.3.0",
    "google-api-python-client>=2.187.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.45",
    "asyncpg>=0.29.0",
    "pyjwt>=2.10.1",
    "greenlet>=3.3.0",
]
